utility.mix.channel / upsampling fixed
upsampling has been broken all this time! oops. should be up 2 not "@up 2".

it does sound better with it on. will eventually make it a settings option. for now it can be disabled by swapping between audio_wrap_up2 and audio_wrap. (or audio_wrap_up4)

not sure how the poly voices are meant to work out the samplerate? a working example is in my tape delay block.

also: finished a first draft of the mix.channel plug. sidechain is a little rough but should do the job. eq presets, sweep and blend all sounding nice.