recording
mixer channel plugs can now record their pre-fader (and pre-console-plugin) signal. this has still been eq'd, sidechained, panned, but the console level fader is lost for now
the main patch needs to send the file name, and the start recording message. it's in config, a proper user interface for that is coming.