optimisation
refactored click zones into a tidier function
could be a lot more aggressive with turning view_changed to false but it's all ready for that in this update.
songs_folder_read only does the more in depth scanning first time and only for the song folder.
ui patchers all now use setfontsize instead of doing it themselves.