fix #399
it was a max quirk (you have to return, wait, start again or it doesn't register the file as closed) AND a benny bug (the code to check if the menu is done was checking for menu[0] and menu[0] wasn't being filled becuase it was a custom prefix block and the custom prefixes had been lost due to first boot. phew. got it.)