fix #472 muting and unmuting use same thresholds
removed optional threshold from mutecontrol_fx

muting and unmuting both use snapshot -> abs -> is > (1 bit at 24b)?

a bit worried that non-audio-rate on the unmute means it'll miss the start of sounds when it unmutes? or delay the start of notes when it unmutes? but i think it'll be fine..