#827 Add offset slider to parameter modulation view
Adds an "offset" slider right below the "gain" slider in the parameter sidebar.
This makes it way quicker to tweak both the depth and centre-point of modulations (like from an LFO) without having to click into the full connection settings. The new slider only shows up for connection types that actually have an offset property.