Update source.sacred.waves.osc.maxpat
the issue was the buffer resize op is (i think) asynchronious so you need to just wait before peeking into it. increasing the wait helped but didn't feel like a real fix.

i swapped how it works: it fills 6 buffers instead of 3 at the start and just tells the gen to switch between the two sets of 3. all the new code is in the wave tables patcher (though i could remove a couple of loadbangs from the main patch)